app/requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
pyodbc==4.0.32
python-dotenv==0.19.2